refactor: shrink the footprint on existing files
Everything the proxy borrowed from core turned out to be re-creatable in
place for less churn than sharing it: parser normalization is four inline
lines (so ParamsParser/parseParams stay in acp.ts untouched), abort-aware
error mapping comes from exporting the existing errorToRequestResult
instead of adding a responder method and rewriting Connection's catch, and
cross-close wiring is two lines in connect() (so linkClosed and the acp.ts
in-memory connect changes revert). Existing-file changes are now: the
export blocks in acp.ts, one function made public in jsonrpc.ts, and
README pointers — no behavior changes outside the new module.
